requirements.txt
python-telegram-bot>=21.8,<22.0
gspread>=6.0.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
